Adding cookies to LoginController and UserSessionData and adding authorization with cookie checks to all the Controller classes

+++ EdiSimpleBackend/src/main/java/com/edi/simplebackend/loans/controller/LoanController.java
@@ -2,10 +2,13 @@ package com.edi.simplebackend.loans.controller;
 
 import com.edi.simplebackend.loans.model.Loan;
 import java.util.List;
+
+import com.edi.simplebackend.login.UserSessionData;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -15,20 +18,31 @@ import org.springframework.web.bind.annotation.*;
 public class LoanController {
 
 	private final LoanService loanService;
+	private final UserSessionData userSessionData;
 
+	@CrossOrigin(origins = "http://localhost:3000", allowCredentials = "true")
 	@PostMapping(params = {"email", "bookId"})
 	public ResponseEntity<Loan> addLoan(
 			@RequestParam final String email,
-			@RequestParam final Long bookId
+			@RequestParam final Long bookId,
+			@RequestHeader(value = "Cookie", required = false) String cookie
 	) {
+		if (!isUserLoggedIn(cookie)) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
 
 		return ResponseEntity.ok(this.loanService.addLoan(email, bookId));
 	}
 
 	@DeleteMapping(params = "loanId")
 	public ResponseEntity<String> deleteLoan(
-			@RequestParam final Long loanId
+			@RequestParam final Long loanId,
+			@RequestHeader(value = "Cookie", required = false) String cookie
 	) {
+
+		if (!isUserLoggedIn(cookie)) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
 
 		loanService.deleteLoan(loanId);
 		return ResponseEntity.ok("Loan " + loanId + " deleted!");
@@ -40,9 +54,14 @@ public class LoanController {
 			@RequestParam final Long bookId,
 			@RequestParam(required = false) final Integer page,
 			@RequestParam(required = false) final Integer size,
-			@RequestParam(required = false) final String sortBy
+			@RequestParam(required = false) final String sortBy,
+			@RequestHeader(value = "Cookie", required = false) String cookie
 
 	) {
+		if (!isUserLoggedIn(cookie)) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
+
 		if (page == null && size == null && sortBy == null) {
 			return ResponseEntity.ok(this.loanService.getLoansByBookId(bookId));
 		} else {
@@ -62,9 +81,14 @@ public class LoanController {
 			@RequestParam final String email,
 			@RequestParam(required = false) final Integer page,
 			@RequestParam(required = false) final Integer size,
-			@RequestParam(required = false) final String sortBy
+			@RequestParam(required = false) final String sortBy,
+			@RequestHeader(value = "Cookie", required = false) String cookie
 
 	) {
+		if (!isUserLoggedIn(cookie)) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
+
 		if (page == null && size == null && sortBy == null) {
 			return ResponseEntity.ok(this.loanService.getLoansByEmail(email));
 		} else {
@@ -80,8 +104,13 @@ public class LoanController {
 
 	@GetMapping(params = "email")
 	public ResponseEntity<List<Object>> getLoansByEmailSimplifiedWithBooks(
-			@RequestParam final String email
+			@RequestParam final String email,
+			@RequestHeader(value = "Cookie", required = false) String cookie
 	) {
+
+		if (!isUserLoggedIn(cookie)) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
 
 		return ResponseEntity.ok(this.loanService.getLoansByEmailWithBooks(email));
 	}
@@ -92,9 +121,14 @@ public class LoanController {
 			@RequestParam final Long userId,
 			@RequestParam(required = false) final Integer page,
 			@RequestParam(required = false) final Integer size,
-			@RequestParam(required = false) final String sortBy
+			@RequestParam(required = false) final String sortBy,
+			@RequestHeader(value = "Cookie", required = false) String cookie
 
 	) {
+		if (!isUserLoggedIn(cookie)) {
+			return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+		}
+
 		if (page == null && size == null && sortBy == null) {
 			return ResponseEntity.ok(this.loanService.getLoansByUserId(userId));
 		} else {
@@ -107,5 +141,22 @@ public class LoanController {
 			return ResponseEntity.ok(this.loanService.getLoansByUserId(userId, pageable));
 		}
 	}
+
+	private boolean isUserLoggedIn(String cookie) {
+		if (cookie != null && cookie.contains("cookieToken")) {
+			// Extract the cookie token from the "cookie" string
+			String[] cookieParts = cookie.split(";");
+			for (String cookiePart : cookieParts) {
+				if (cookiePart.trim().startsWith("cookieToken=")) {
+					String token = cookiePart.trim().substring("cookieToken=".length());
+					return token.equals(userSessionData.getCookieToken());
+				}
+			}
+		}
+		return false;
+	}
+
+
+
 
 }
